Disabled SSL support through distlib.
Pip should work even when no SSL is available, but currently it fails due to an
import error. This is a quick hack to disable SSL support and use pip in a
situation where no SSL is available.

Note that the rest of pip can handle missing SSL support and even warns about
it.